refactor and change file structure
