+++ configs/config.yaml
@@ -15,12 +15,15 @@ optimizer:
 
 data:
   seg_data_path: /workspace/data/segmentation_dataset_512/ # segmentation dataset path
-  gait_seg_data_path: /workspace/data/segmentation_dataset_512/json # defined gait cycle json path 
-  gait_seg_index_data_path: /workspace/data/seg_gait_index_mapping/ # training mapping path
+  gait_seg_data_path: /workspace/data/segmentation_dataset_512/json_mix # defined gait cycle json path. This path uesd be gait cycle defined dataset. 
+  gait_seg_index_data_path: /workspace/data/seg_gait_index_mapping/ # training mapping path, this used for cross validation, with different class number.
 
   num_workers: 4
   img_size: 224
-  sampling: "over" # over, under, none
+  sampling: "under" # over, under, none
+
+  gait_cycle_batch_size: 1 # this used for gait cycle defined method. 
+  default_batch_size: 32  # this used for default method, without gait cycle defined method.
 
 model:
   model: "resnet"
@@ -38,7 +41,6 @@ YOLO:
 train:
   # Training config
   max_epochs: 50 # numer of epochs of training
-  batch_size: 1 # batch size for the dataloader
 
   # used for val 
   clip_duration: 0.5 # clip duration for the video
