+++ configs/prepare_dataset.yaml
@@ -16,15 +16,15 @@ YOLO:
 # step1: from raw video to extract video.
 # use YOLOv8 detection with tracking.
 extract_dataset:
-  data_path: ""
-  save_path: "/workspace/data/segmentation_dataset_512"
+  data_path: "" # this is raw data path.
+  save_path: "/workspace/data/segmentation_dataset_512" # this is extracted video path.
 
 # step2: from extracted video to define the gait cycle.
 gait_dataset:
   data_path: "/workspace/data/segmentation_dataset_512" # meta four view dataset path, with human annotation.
-  save_path: "/workspace/data/segmentation_gait_dataset_512" # convert the four view dataset into a format of the annotation file.
+  save_path: "/workspace/data/seg_gait_cycle_index" # convert the four view dataset into a format of the annotation file.
   gait_cycle_split: 2 # define the gait cycle split number.  
   define_part: "left_ankle"
 
-log_path: "${gait_dataset.data_path}/json/log/"
-method: 'bbox' # bbox or OF
+log_path: "${gait_dataset.data_path}/json_${method}/log/"
+method: 'bbox' # bbox or pose or mix
