--- prepare/prepare_annotation/labels_to_id.csv
@@ -1,5 +1,0 @@
-id,name
-0,ASD
-1,DHS
-2,LCS
-3,Normal
