+++ project/dataloader/data_loader.py
@@ -15,6 +15,9 @@ HISTORY:
 Date 	By 	Comments
 ------------------------------------------------
 
+25-03-2024	Kaixu Chen	change batch size for train/val dataloader. Now set bs=1 for gait cycle dataset, set bs=32 for default dataset (without gait cycle).
+Because in experiment, I found bs=1 will have large loss when train. Maybe also need gait cycle datset in val/test dataloader?
+
 22-03-2024	Kaixu Chen	add different class num mapping dict. In collate_fn, re-mapping the label from .json disease key.
 
 """
@@ -26,20 +29,12 @@ from pathlib import Path
 
 from torchvision.transforms import (
     Compose,
-    Lambda,
-    RandomCrop,
     Resize,
-    RandomHorizontalFlip,
-    ToTensor,
 )
 from pytorchvideo.transforms import (
     ApplyTransformToKey,
-    Normalize,
-    RandomShortSideScale,
-    ShortSideScale,
     UniformTemporalSubsample,
     Div255,
-    create_video_transform,
 )
 
 from typing import Any, Callable, Dict, Optional, Type
@@ -53,22 +48,11 @@ from pytorchvideo.data.labeled_video_dataset import labeled_video_dataset
 from gait_video_dataset import labeled_gait_video_dataset
 
 disease_to_num_mapping_Dict: Dict = {
-    2: {
-        "ASD": 0,
-        "non-ASD": 1
-    },
-    3: {
-        "ASD": 0,
-        "DHS": 1,
-        "LCS_HipOA": 2
-    },
-    4: {
-        "ASD": 0,
-        "DHS": 1,
-        "LCS_HipOA": 2,
-        "normal": 3
-    }
+    2: {"ASD": 0, "non-ASD": 1},
+    3: {"ASD": 0, "DHS": 1, "LCS_HipOA": 2},
+    4: {"ASD": 0, "DHS": 1, "LCS_HipOA": 2, "normal": 3},
 }
+
 
 class WalkDataModule(LightningDataModule):
     def __init__(self, opt, dataset_idx: Dict = None):
@@ -78,7 +62,10 @@ class WalkDataModule(LightningDataModule):
         self._gait_seg_path = opt.data.gait_seg_data_path
         self.gait_cycle = opt.train.gait_cycle
 
-        self._BATCH_SIZE = opt.train.batch_size
+        # ? 感觉batch size对最后的结果有影响，所以分开使用不同的batch size
+        self._gait_cycle_batch_size = opt.data.gait_cycle_batch_size
+        self._default_batch_size = opt.data.default_batch_size
+
         self._NUM_WORKERS = opt.data.num_workers
         self._IMG_SIZE = opt.data.img_size
 
@@ -103,13 +90,16 @@ class WalkDataModule(LightningDataModule):
         self.train_video_transform = Compose(
             [
                 ApplyTransformToKey(
-                    key="video", 
+                    key="video",
                     transform=Compose(
                         [
                             Div255(),
                             Resize(size=[self._IMG_SIZE, self._IMG_SIZE]),
-                            UniformTemporalSubsample(self.uniform_temporal_subsample_num),
-                        ])
+                            UniformTemporalSubsample(
+                                self.uniform_temporal_subsample_num
+                            ),
+                        ]
+                    ),
                 ),
             ]
         )
@@ -137,28 +127,7 @@ class WalkDataModule(LightningDataModule):
         so we directly use the pytorchvideo API to load the video.
         AKA, use whole video to validate the model.
         """
-
-        # temp_val_path = Path(self._seg_path) / "val_temp"
-        # val_idx = self._dataset_idx[1]
-
-        # shutil.rmtree(temp_val_path, ignore_errors=True)
-
-        # for path in val_idx:
-        #     with open(path) as f:
-        #         file_info_dict = json.load(f)
-
-        #     video_name = file_info_dict["video_name"]
-        #     video_path = file_info_dict["video_path"]
-        #     video_disease = file_info_dict["disease"]
-
-        #     if not (temp_val_path / video_disease).exists():
-        #         (temp_val_path / video_disease).mkdir(parents=True, exist_ok=False)
-
-        #     shutil.copy(
-        #         video_path, temp_val_path / video_disease / (video_name + ".mp4")
-        #     )
-
-        # self.temp_val_path = temp_val_path
+        ...
 
     def setup(self, stage: Optional[str] = None) -> None:
         """
@@ -181,10 +150,10 @@ class WalkDataModule(LightningDataModule):
                 # * labeled dataset, where first define the giat cycle, and from .json file to load the video.
                 # * here only need dataset idx, mean first split the dataset, and then load the video.
                 self.train_gait_dataset = labeled_gait_video_dataset(
-                    gait_cycle = self.gait_cycle,
+                    gait_cycle=self.gait_cycle,
                     dataset_idx=self._dataset_idx[0],  # [train, val]
                     transform=self.train_mapping_transform,
-                    temporal_mix=self._temporal_mix
+                    temporal_mix=self._temporal_mix,
                 )
 
         if stage in ("fit", "validate", None):
@@ -218,7 +187,7 @@ class WalkDataModule(LightningDataModule):
         batch_label = []
         batch_video = []
 
-        # mapping label 
+        # mapping label
 
         for i in batch:
             # logging.info(i['video'].shape)
@@ -227,13 +196,17 @@ class WalkDataModule(LightningDataModule):
 
             batch_video.append(i["video"])
             for _ in range(gait_num):
-                
+
                 if disease in disease_to_num_mapping_Dict[self._class_num].keys():
 
-                    batch_label.append(disease_to_num_mapping_Dict[self._class_num][disease])
+                    batch_label.append(
+                        disease_to_num_mapping_Dict[self._class_num][disease]
+                    )
                 else:
                     # * if the disease not in the mapping dict, then set the label to non-ASD.
-                    batch_label.append(disease_to_num_mapping_Dict[self._class_num]['non-ASD'])
+                    batch_label.append(
+                        disease_to_num_mapping_Dict[self._class_num]["non-ASD"]
+                    )
 
         # video, b, c, t, h, w, which include the video frame from sample info
         # label, b, which include the video frame from sample info
@@ -253,7 +226,7 @@ class WalkDataModule(LightningDataModule):
         if self.gait_cycle == -1:
             train_data_loader = DataLoader(
                 self.train_gait_dataset,
-                batch_size=self._BATCH_SIZE,
+                batch_size=self._default_batch_size,
                 num_workers=self._NUM_WORKERS,
                 pin_memory=True,
                 shuffle=False,
@@ -262,14 +235,14 @@ class WalkDataModule(LightningDataModule):
 
         else:
             train_data_loader = DataLoader(
-            self.train_gait_dataset,
-            batch_size=self._BATCH_SIZE,
-            num_workers=self._NUM_WORKERS,
-            pin_memory=True,
-            shuffle=True,
-            drop_last=True,
-            collate_fn=self.collate_fn,
-        )
+                self.train_gait_dataset,
+                batch_size=self._gait_cycle_batch_size,
+                num_workers=self._NUM_WORKERS,
+                pin_memory=True,
+                shuffle=True,
+                drop_last=True,
+                collate_fn=self.collate_fn,
+            )
 
         return train_data_loader
 
@@ -282,23 +255,23 @@ class WalkDataModule(LightningDataModule):
 
         return DataLoader(
             self.val_gait_dataset,
-            batch_size=self._BATCH_SIZE,
+            batch_size=self._default_batch_size,
             num_workers=self._NUM_WORKERS,
             pin_memory=True,
             shuffle=False,
             drop_last=True,
         )
-    
+
     def test_dataloader(self) -> DataLoader:
         """
         create the Walk train partition from the list of video labels
         in directory and subdirectory. Add transform that subsamples and
         normalizes the video before applying the scale, crop and flip augmentations.
         """
-            
+
         return DataLoader(
             self.test_gait_dataset,
-            batch_size=self._BATCH_SIZE,
+            batch_size=self._default_batch_size,
             num_workers=self._NUM_WORKERS,
             pin_memory=True,
             shuffle=False,
